Updated EmbPattern
Updated the EmbPattern to use an internal datastructure that still uses floats[] for the points but adds in int data[] for each pair of points. Allows stitchblocks to be references to that data. Adds in some powerful iterators. Added a share_fragment that shares all the writing embroidery file formats. Changed the drawer fragment to be both colors and stitches (reflecting the different EmbPattern. Added in an internal file type to do actual saving of the current pattern.